fix(tabs): quick-fork lands as a pane-stack tab, not a new window tab
Repo-owner caught this live testing: the shipped quick-fork (PR #2725/#2727/#2735)
opened forks in a brand-new top-level window tab, triggered from the window
tab-strip's own right-click menu. That was wrong on two counts: (1) forking is a
per-agent action, and the codebase already has a per-agent right-click surface
for exactly that (AgentViewModel.getBodyContextMenuItems, home to "Agent
History") -- the window tab-strip's menu is ambiguous once a tab can hold more
than one pane; (2) launchAgentDefinition's own targetBlockId param already
existed (since 2026-07-20, a month before this feature) specifically for "the
fork-tab-strip + action" -- i.e. pushing the fork onto the SAME pane's own
blockStack, per SPEC_AGENT_PANE_FORKS_AND_AUX_PINS_2026_06_15.md's own
(never-wired) design.

Moves quickForkTabToNewTab (frontend/app/tab/quick-fork.ts) to quickForkAgent
(frontend/app/view/agent/quick-fork.ts): forks now push a new block onto the
source pane's own stack via pushBlockOntoStack -- the same primitive
open-history-tab.ts's "Agent History" entry and the pane tab strip's own "+"
already use -- instead of creating a new WorkspaceService tab. Triggered from
getBodyContextMenuItems right next to "Agent History". Removes the tab-strip
context-menu entries, the Cmd:Shift:t keybinding (its target no longer makes
sense post-correction), and launchAgentDefinition's now-dead targetTabId param.
