+++ .changesets/1787436301-fix-tabs-move-quick-fork-from-the-window-tab-strip-into-the--1mqa.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+fix(tabs): move quick-fork from the window tab-strip into the per-agent pane, landing forks as a sibling pane tab
